+++ doc.go
@@ -2,6 +2,6 @@
 // output destinations, and message templates.
 //
 // The loggo package allows you to create a Logger with a specified log level
-// threshold and various options to customize its behavior. You can log messages
+// Threshold and various options to customize its behavior. You can log messages
 // at different levels such as Debug, Info, Warn, Error, and Fatal.
 package loggo